Add most useful items to the System menu
Not all menu items are added, but only "most useful items", like New Tab,
Close Tab, Edit, About.

The items are added to the system menu with slightly different hot keys for
two reasons:
- to not collapse with the default items that are already there (like
  Mi&nimize vs. &New tab);
- to mimic cmd's extra items (e.g. Alt+Space,E,P executes Paste;
  Alt+Space,P opens Properties, which are really Settings).

The items are copied from the template IDR_SYSTEM_MENU.

The New Tab sub-menu is not re-created each time a tab is added or removed
to not go through troubles of finding the item in System menu.