Delete Zone.Identifier to unblock contents of the help file
By default, the Help will not be displayed if the CHM file was downloaded
from the Internet, thus contains Zone.Identifier, which is documented at
http://msdn.microsoft.com/en-us/library/ff469212(PROT.10).aspx

The solution will work in most cases, but not if the administrator unzipped
Console for all users to use it. The proper solution is to have a setup.